backtester/requirements.txt
python-binance==0.7.11
pandas==1.2.0
arrow==0.17.0
